requirements.txt
adal==1.2.7
altair==5.5.0
applicationinsights==0.11.10
attrs==24.2.0
azure-ai-ml==1.22.4
azure-common==1.1.28
azure-core==1.32.0
azure-graphrbac==0.61.2
azure-identity==1.19.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.1
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
azure-storage-file-share==12.20.0
azureml-core==1.0.23
azureml-pipeline==1.0.23
azureml-pipeline-core==1.0.23
azureml-pipeline-steps==1.0.23
azureml-sdk==1.0.23
azureml-telemetry==1.0.23
azureml-train==1.0.23
azureml-train-core==1.0.23
azureml-train-restclients-hyperdrive==1.0.23
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
cryptography==44.0.0
docker==7.1.0
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
huggingface-hub==0.26.5
idna==3.10
isodate==0.7.2
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpickle==4.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
narwhals==1.18.4
ndg-httpsclient==0.5.1
networkx==3.4.2
numpy==2.2.0
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
pathspec==0.12.1
pillow==11.0.0
portalocker==2.10.1
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydash==8.0.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.15.89
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
SecretStorage==3.3.3
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
streamlit==1.41.1
strictyaml==1.7.3
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.45.1
